Refactor provider registry to domain layer
